+++ kibot/misc.py
@@ -173,6 +173,7 @@ W_NOKIVER = '(W052) '
 W_EXTNAME = '(W053) '
 W_TIMEOUT = '(W054) '
 W_MUSTBEINT = '(W055) '
+W_NOOUTPUTS = '(W056) '
 
 
 class Rect(object):